requirements.txt
Django
djangorestframework-simplejwt
djangorestframework
django-cors-headers
sendgrid
Pillow
psycopg2
twilio
stripe